+++ tests/testthat/test-gather.R
@@ -1,0 +1,11 @@
+context("Gather")
+
+test_that("gather all columns when ... is empty", {
+  df <- data.frame(
+    x = 1:5,
+    y = 6:10
+  )
+  out <- gather(df, key, val)
+  expect_equal(nrow(out), 10)
+  expect_equal(names(out), c("key", "val"))
+})